+++ tests/CSharpLanguageServer.Tests/Fixtures/genericProject/Project/NameSuggestionCompletionTests.cs
@@ -1,0 +1,7 @@
+class NameSuggestionCompletionTests
+{
+    public void MethodA()
+    {
+        System.Text.StringBuilder 
+    }
+}